Modified settings logic to load connections without crashing

+++ src/main_window.hpp
@@ -180,15 +180,15 @@ private:
   inline void saveDAQSettings(QSettings& userprefs);
   inline void loadDAQSettings(
       QSettings& userprefs,
-      std::unordered_map<size_t, IO::Block*>& block_cache);
+      std::unordered_map<int, IO::Block*>& block_cache);
   inline void saveWidgetSettings(QSettings& userprefs);
   inline void loadWidgetSettings(
       QSettings& userprefs,
-      std::unordered_map<size_t, IO::Block*>& block_cache);
+      std::unordered_map<int, IO::Block*>& block_cache);
   inline void saveConnectionSettings(QSettings& userprefs);
   inline void loadConnectionSettings(
       QSettings& userprefs,
-      std::unordered_map<size_t, IO::Block*>& block_cache);
+      std::unordered_map<int, IO::Block*>& block_cache);
   Event::Manager* event_manager;
   QMdiArea* mdiArea = nullptr;
   QList<QMdiSubWindow*> subWindows;
